+++ NEWS.md
@@ -3,6 +3,8 @@
 * Add run_reason argument to `benchmark_results`
 * deprecated `benchmark` in favour of `benchmark_results`
 * Hard dependency on httr 0.2.2 for escaping capabilities
+* Enforce only 5 run_ids for `benchmark_results`
+* Assert that `benchmark_results` can only use one of run_id, batch_id or run_reason
 
 # conbenchcoms 0.0.1